requirements.txt
pyserial>=3.2.1
pyserial-asyncio>=0.4
IntelHex>=2.1
crcmod>=1.7
voluptuous==0.10.5
